Update README.md
"This README provides detailed instructions and necessary information for installing, activating, and customizing a custom WordPress theme. It includes clear guidelines for cloning the repository, setting up WordPress and the database, and activating the theme via the WordPress admin panel. Additionally, it outlines how to contribute to the project and offers contact information for any questions or collaboration. It's an essential document for users looking to utilize and personalize the theme on their WordPress site."